# wa-worker/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "WA worker with rest api",
        "title": "WA-WORKER",
        "contact": {
            "name": "ymkp",
            "url": "https://github.com/ymkp",
            "email": "banterbedlam@gmail.com"
        },
        "version": "0.x"
    },
    "host": "127.0.0.1:3000",
    "paths": {
...
        "/api/v1/whatsapp/group": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Joined Groups Information from WhatsApp",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Group"
                ],
                "summary": "Get Joined Groups Information",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/login": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get QR Code for WhatsApp Multi-Device Login",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "text/html"
                ],
                "tags": [
                    "WhatsApp Authentication"
                ],
                "summary": "Generate QR Code for WhatsApp Multi-Device Login",
                "parameters": [
                    {
                        "enum": [
                            "html",
                            "json"
                        ],
                        "type": "string",
                        "default": "html",
                        "description": "Change Output Format in HTML or JSON",
                        "name": "output",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Make Device Logout from WhatsApp Multi-Device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Authentication"
                ],
                "summary": "Logout Device from WhatsApp Multi-Device",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/registered": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check WhatsApp Personal ID is Registered",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Authentication"
                ],
                "summary": "Check If WhatsApp Personal ID is Registered",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WhatsApp Personal ID to Check",
                        "name": "msisdn",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/audio": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Audio Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Audio Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Audio File",
                        "name": "audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/contact": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Contact Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Contact Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact Phone",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/document": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Document Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Document Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Document File",
                        "name": "document",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Image Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Image Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Caption Image Message",
                        "name": "caption",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image File",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Is View Once",
                        "name": "viewonce",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/link": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Link Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Link Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Link Caption",
                        "name": "caption",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Link URL",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/location": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Location Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Location Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Location Latitude",
                        "name": "latitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Location Longitude",
                        "name": "longitude",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/sticker": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Sticker Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Sticker Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Sticker File",
                        "name": "sticker",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/whatsapp/send/text": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Text Message to Spesific WhatsApp Personal ID or Group ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Message"
                ],
                "summary": "Send Text Message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination WhatsApp Personal ID or Group ID",
                        "name": "msisdn",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text Message",
                        "name": "message",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}